+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- no user facing changes
+
 ## 3.7.2 - 2024-09-17
 
 - no user facing changes